requirements.txt
googletrans==4.0.0-rc1
langdetect
openai
flask
python-docx
tiktoken
markdown2